% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d2-utils.R
\name{d2-utils}
\alias{d2-utils}
\alias{d2_which}
\alias{d2_version}
\alias{d2_themes}
\alias{d2_layout}
\title{Basic D2 utilities}
\usage{
d2_which()

d2_version()

d2_themes()

d2_layout(layout = NULL)
}
\arguments{
\item{layout}{Layout name. If supplied, display long help for a particular layout engine,
including its configuration options.}
}
\description{
Basic D2 utilities

\itemize{
\item \code{\link[=d2_which]{d2_which()}}: List path for D2
}

\itemize{
\item \code{\link[=d2_version]{d2_version()}}: List the installed version of D2
}

\itemize{
\item \code{\link[=d2_themes]{d2_themes()}}: List available D2 themes
}

\itemize{
\item \code{\link[=d2_layout]{d2_layout()}}: Lists available layout engine options with short help
}
}
