% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagram.R
\name{d2_container}
\alias{d2_container}
\title{Create a D2 container or map}
\usage{
d2_container(
  ...,
  lines = NULL,
  id = NULL,
  class = NULL,
  name = NULL,
  label = NULL,
  shape = NULL,
  width = NULL,
  height = NULL,
  icon = NULL,
  style = d2_style(),
  collapse = NULL
)
}
\arguments{
\item{...}{Diagram text.}

\item{lines}{Optional diagram text. If supplied \code{lines} is recycled to match
the length of \code{id}.}

\item{id}{D2 diagram map identifier. All other container or map attributes
are recycled to match the length of \code{id} using \code{\link[vctrs:vec_recycle]{vctrs::vec_recycle_common()}}}

\item{class}{Diagram class}

\item{name}{Diagram container name}

\item{label}{Diagram label}

\item{shape}{Diagram shape}

\item{width, height}{Diagram width and height (in pixels?)}

\item{icon}{Diagram icon}

\item{style}{A string or results from the \code{\link[=d2_style]{d2_style()}} helper function.}

\item{collapse}{If \code{collapse = "\\n"}, return a string with D2 diagram code for a
container. If \code{collapse = NULL} (default), return a character vector with the same
length as id.}
}
\value{
A string or character vector with D2 diagram code.
}
\description{
A container or map is a group of diagram elements. Some special chart types,
such as the SQL Table created by \code{\link[=d2_sql_table]{d2_sql_table()}} must be set up as a
container. Using containers can also help set element properties including
the class, label, shape and dimensions (width and height) for a set of
diagram elements.
}
\examples{
d2_container(
  "team1 <-> team2", "team2 <-> team3", "team3 <-> team1",
  id = c("round1", "round2")
)

d2_container(
  lines = c("shark -> fish: eats", "lion -> gazelle: eats"),
  id = c("ocean", "savannah")
)

}
