% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagram.R
\name{d2_diagram}
\alias{d2_diagram}
\title{Create a D2 diagram from a named vector and connectors}
\usage{
d2_diagram(
  lines = NULL,
  ...,
  connector = NULL,
  id = NULL,
  direction = getOption("d2r.direction"),
  import = NULL,
  collapse = NULL
)
}
\arguments{
\item{lines}{A list or character vector. Named elements are connected using
the supplied \code{connector} value. Unnamed elements are kept as is and assumed
to be valid D2 code.}

\item{...}{Additional elements for diagram included after initial lines.}

\item{connector}{Default to \code{NULL} which uses \code{"->"} to connect the name of
any shape to the value of any shape. Must be vector of:
--, ->, <-, and <->. Vector is recycled to
match length of named shapes.}

\item{id}{Identifier for map container for diagram.}

\item{direction}{Optional direction for diagram. Must be one of:
up, down, right, and left.}

\item{import}{File name or names to append as an import to the top of the
diagram. If named, the standard import method using the name of the import
as the name of the imported element. If import is unnamed, the partial
import method is used: https://d2lang.com/tour/imports#partial-imports See:
\url{https://d2lang.com/tour/imports}}

\item{collapse}{an optional character string to separate the results.  Not
    \code{\link[base]{NA_character_}}.  When \code{collapse} is a string,
    the result is always a string (\code{\link[base]{character}} of length 1).}
}
\description{
Note: this is an initial experimental implementation of diagram building
and may change or be deprecated in the future.
}
\examples{
d2_diagram(c("x" = "y"))

d2_diagram(c("a" = "b"), connector = "<-", direction = "up")

d2_diagram(c("start" = "end"), connector = "<->", import = "imported.d2")

}
