% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{d2_ggplot}
\alias{d2_ggplot}
\title{Plot a D2 diagram as a ggplot2 using \code{magick::image_ggplot()}}
\usage{
d2_ggplot(
  x,
  ...,
  density = 150,
  width = NULL,
  height = NULL,
  interpolate = FALSE,
  arg = caller_arg(x),
  call = caller_env()
)
}
\arguments{
\item{x}{A character vector of diagram text or a file path for a D2 diagram
file to render as a diagram with \code{\link[=d2_render]{d2_render()}} before plotting. A file path
for a rendered D2 diagram file one using a svg, pdf, png, or gif file
extension is also allowed.}

\item{...}{Additional parameters passed to \code{\link[=d2_diagram]{d2_diagram()}}.}

\item{density}{Resolution to render pdf passed to \code{\link[magick:editing]{magick::image_read_pdf()}}.
Default \code{150}.}

\item{width}{in pixels}

\item{height}{in pixels}

\item{interpolate}{passed to \link[ggplot2:annotation_raster]{ggplot2::annotation_raster}}

\item{arg}{An argument name as a string. This argument
will be mentioned in error messages as the input that is at the
origin of a problem.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
\code{\link[=d2_ggplot]{d2_ggplot()}} renders a D2 diagram input character vector or file path (to a
diagram file or rendered diagram) and plots the image with
\code{\link[magick:image_ggplot]{magick::image_ggplot()}}.
}
