% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagram.R
\name{d2_map}
\alias{d2_map}
\title{Create a D2 map}
\usage{
d2_map(..., .id = NULL, .class = NULL, .label = NULL, .shape = NULL)
}
\arguments{
\item{...}{Text lines for diagram.}

\item{.id}{D2 diagram map identifier.}

\item{.class}{Diagram map class}

\item{.label}{Diagram map label}

\item{.shape}{Diagram map shape}
}
\description{
A map is a group within the diagram. This is required for \code{\link[=d2_sql_table]{d2_sql_table()}}
but can to set properties such as the class, label, or shape for a grouped
set of diagram elements.
}
\examples{
d2_map(
  "principal -> teachers",
  "teachers -> students",
  .id = "school",
  .shape = "circle"
)

}
