% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d2-utils.R
\name{d2_read}
\alias{d2_read}
\alias{d2_write}
\alias{d2_fmt}
\title{Read, write, and format D2 files}
\usage{
d2_read(file)

d2_write(data, file = tempfile(fileext = fileext), fileext = "d2")

d2_fmt(file)
}
\arguments{
\item{file}{File name to write to disk.}

\item{data}{A set of lines or a file path to a d2 file.}

\item{fileext}{File extension to use for writing data to disk.}
}
\description{
Read, write, and format D2 files
}
