% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special.R
\name{d2_sql_table}
\alias{d2_sql_table}
\title{Create a D2 SQL Table Diagram}
\usage{
d2_sql_table(
  data = NULL,
  visibility = NULL,
  types = vec_ptype_abbr,
  .id = caller_arg(data),
  ...
)
}
\arguments{
\item{data}{A data frame used for column names and types.}

\item{visibility}{One or more of: , +, -, #. Length is
recycled to match the number of columns in \code{data}.}

\item{types}{A character vector matching the number of columns in data or a
function used to return the character vector. Defaults to
\link[vctrs:vec_ptype_full]{vctrs::vec_ptype_abbr}. Other recommended options are
\link[vctrs:vec_ptype_full]{vctrs::vec_ptype_full} or \link[base:typeof]{base::typeof}. If \code{types} is a named character
vector, the column names of \code{data} are ignored the names of types used
instead.}

\item{.id}{Table ID passed to \code{\link[=d2_map]{d2_map()}}. Defaults to
\code{rlang::caller_arg(data)}}

\item{...}{Unused at present.}
}
\description{
Create a D2 SQL Table Diagram
}
