% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{d2_style}
\alias{d2_style}
\title{Create a D2 style}
\usage{
d2_style(
  ...,
  inline = FALSE,
  level = c("shape", "connector", "root"),
  after = "\\n"
)
}
\arguments{
\item{...}{Named attributes for style. Underscores in parameter names are
substituted for dashes to match D2 style, e.g. \code{"stroke_dash"} is converted
to \code{"stroke-dash"}.}

\item{inline}{If \code{TRUE} or a string value , return a string instead of an array. Errors in either case
and more than a single parameter is supplied. If string, the string is
assumed to be the id for a shape, connector, or container/map and appended
before the "style" keyword.}

\item{level}{String. "shape", "connector", or "root". If "root", limit style to
supported attributes for root-level style. See
\url{https://d2lang.com/tour/style#root} for more information.}

\item{after}{String after each line of the style specification. Defaults to
\code{"\\n"}}
}
\description{
Learn more: \url{https://d2lang.com/tour/style}
}
\examples{
d2_style(stroke = "#f4a261", inline = TRUE)

d2_style(fill = "green", inline = "tree")

d2_style(animated = TRUE, stroke_dash = 3, level = "connector")

}
