% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d2r-package.R
\docType{package}
\name{d2r-package}
\alias{d2r}
\alias{d2r-package}
\title{d2r: Create Diagrams with D2}
\description{
Build, read, write, and render diagrams using the D2 syntax.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/elipousson/d2r}
  \item Report bugs at \url{https://github.com/elipousson/d2r/issues}
}

}
\author{
\strong{Maintainer}: Eli Pousson \email{eli.pousson@gmail.com} (\href{https://orcid.org/0000-0001-8280-1706}{ORCID}) [copyright holder]

}
\keyword{internal}
