% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d2-utils.R
\name{is_d2_file}
\alias{is_d2_file}
\title{Does file use a d2 extension?}
\usage{
is_d2_file(file, fileext = "d2")
}
\arguments{
\item{file}{File name(s) to test.}

\item{fileext}{File extension(s) to test for.}
}
\description{
\code{\link[=is_d2_file]{is_d2_file()}} tests for a \code{".d2"} file extension. A d2 file extension is not
required in all cases but most functions in this package expect a d2
extension by default.
}
