% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{themes_d2}
\alias{themes_d2}
\title{D2 Themes}
\format{
A length 19 double vector.
}
\source{
\url{https://d2lang.com/tour/themes/}
}
\usage{
themes_d2
}
\description{
A vector of standard D2 theme codes named with theme names.
}
\keyword{datasets}
