% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{d2-utils}
\alias{d2-utils}
\alias{d2_version}
\alias{d2_themes}
\alias{d2_layout}
\alias{is_d2_file}
\title{Basic D2 utilities}
\usage{
d2_version()

d2_themes()

d2_layout(layout = NULL)

is_d2_file(file, fileext = "d2")
}
\arguments{
\item{layout}{Layout name. If supplied, display long help for a particular layout engine,
including its configuration options.}
}
\description{
Basic D2 utilities

\itemize{
\item \code{\link[=d2_version]{d2_version()}}: List the version of D2 (if installed)
}

\itemize{
\item \code{\link[=d2_themes]{d2_themes()}}: List available D2 themes
}

\itemize{
\item \code{\link[=d2_layout]{d2_layout()}}: Lists available layout engine options with short help
}

\itemize{
\item \code{\link[=is_d2_file]{is_d2_file()}}: Test for a \code{".d2"} file extension
}
}
