% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagram.R
\name{d2_diagram}
\alias{d2_diagram}
\title{Format a vectors of shapes and connectors as a D2 diagram}
\usage{
d2_diagram(
  shapes,
  connector = NULL,
  class = NULL,
  direction = NULL,
  import = NULL,
  ...
)
}
\description{
Format a vectors of shapes and connectors as a D2 diagram
}
