% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{d2_render}
\alias{d2_render}
\title{Render a D2 file}
\usage{
d2_render(
  input,
  output = NULL,
  fileext = "svg",
  layout = "elk",
  theme = NULL,
  sketch = NULL,
  pad = NULL,
  animate_interval = NULL,
  ...,
  overwrite = TRUE
)
}
\arguments{
\item{input}{Required input file with \code{"d2"} file extension.}

\item{...}{Additional input flags. Optional character vector.}
}
\description{
Render a D2 file
}
