% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{keys_d2}
\alias{keys_d2}
\title{D2 Syntax Keywords}
\format{
A length 9 named list.
}
\source{
\url{https://d2lang.com/}
}
\usage{
keys_d2
}
\description{
An incomplete list of standard D2 syntax keywords for reference and internal
validation. Named elements include: layout, theme, fileext, connector, arrowhead, direction, style, visibility, and grid.
}
\keyword{datasets}
